Spike trap werkt + inventory werkt weer

+++ Ketopia/Assets/Nobe/CustomEditors/Interfaces/IDamagable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamagable 
 {
-    public void IDamagable(int dmgDone, NodeType typeUsed, int toolStrength);
+    public void IDamagable(float dmgDone, NodeType typeUsed, int toolStrength);
 }
